-Merge changes from 3.5 trunk -Update testregression for latest kdelibs changes -Update to support latest Qt4.1 snapshots.
svn path=/trunk/KDE/kdelibs/; revision=480061

+++ khtml/rendering/render_br.h
@@ -49,7 +49,7 @@ public:
     // overrides
     virtual void calcMinMaxWidth() {}
     virtual short minWidth() const { return 0; }
-    virtual short maxWidth() const { return 0; }
+    virtual int maxWidth() const { return 0; }
 
     virtual FindSelectionResult checkSelectionPoint( int _x, int _y, int _tx, int _ty,
                                                      DOM::NodeImpl*& node, int & offset,
